requirements.txt
basic_colormath==0.5.0
jupyter==1.1.1
matplotlib==3.10.3
numpy==2.3.1
distinctipy==1.3.4
scipy==1.16.0
scikit_learn==1.6.1
Pillow==11.3.0

# dev
ruff==0.12.4
